音乐播放器 播放列表

+++ app/src/main/java/com/wgllss/ssmusic/core/units/UUIDHelp.kt
@@ -5,11 +5,11 @@ import java.util.*
 
 object UUIDHelp {
 
-    fun getMusicUUID(musicBean: MusicBean): Int {
+    fun getMusicUUID(musicBean: MusicBean): Long {
         val sb = StringBuilder()
         musicBean?.run {
             sb.append(title).append(author).append(url).append(pic).append(toString())
         }
-        return Math.abs(UUID(sb.toString().hashCode().toLong(), musicBean.toString().hashCode().toLong()).toString().replace("-", "").hashCode())
+        return Math.abs(UUID(sb.toString().hashCode().toLong(), musicBean.toString().hashCode().toLong()).toString().replace("-", "").hashCode()).toLong()
     }
 }
